requirements.txt
pandas==1.4.3
openpyxl==3.0.10
tabulate==0.8.10
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0